+++ Source/DelayEngine.h
@@ -3,6 +3,10 @@
 #include "Parameters.h"
 #include "ChorusEngine.h"
 #include "StereoDelay.h"
+#include "FilterEngine.h"
+#if JUCE_DEBUG
+#include "LFOScopeBuffer.h"
+#endif
 
 using Filter = juce::dsp::StateVariableTPTFilter<float>;
 
@@ -19,10 +23,6 @@ public:
 
 	void update(const Parameters& params);
 	void processSample(const float& inL, const float& inR, float& outL, float& outR, const Parameters& params);
-	void setLowCut(const Parameters& params);
-	void setHighCut(const Parameters& params);
-	void setFilterFreq(const float freq, float& currentFreq, Filter& filter);
-	void setFilterQ(const float q, float& currentQ, Filter& filter);
 	void setDelayTimes(const float delayInMsL, const float delayInMsR);
 	void setFeedbackLevel(const float value) { m_feedbackLevel = value; }
 	void setMixLevel(const float value) { m_mixLevel = value; }
@@ -30,13 +30,19 @@ public:
 	void setWidthLevel(const float value) { m_widthLevel = value; }
 	void setOffsetMs(const float value) { m_offsetMs = value * .5f; }
 	void setDelayMode(const int modeIndex);
+	void updateSyncedModulators(const double ppqPosition);
+
+	std::atomic<float>& getLowCutLfoValue()  { return m_filterEngine.getLowCutLfoValue(); }
+	std::atomic<float>& getHighCutLfoValue() { return m_filterEngine.getHighCutLfoValue(); }
+
+#if JUCE_DEBUG
+	LFOScopeBuffer& getLFOScopeBuffer() { return m_filterEngine.m_scopeBuffer; }
+#endif
 
 private:
-	//juce::dsp::DelayLine<float, juce::dsp::DelayLineInterpolationTypes::Linear> m_delayLine;
 	StereoDelay m_stereoDelay;
 	DelayMode m_delayMode = DelayMode::Stereo;
-	Filter m_lowCutFilter;
-	Filter m_highCutFilter;
+	FilterEngine m_filterEngine;
 	Filter m_feedbackHighpass;
 	juce::dsp::Compressor<float> m_feedbackCompressor;
 	ChorusEngine m_chorusEngine;
@@ -57,10 +63,6 @@ private:
 	float m_gainLevel = 1.0f;
 	float m_widthLevel = 1.0f;
 	float m_offsetMs = 0.0f;
-	float m_lowCutFreq = -1.0f;
-	float m_lowCutQ = -1.0f;
-	float m_highCutFreq = -1.0f;
-	float m_highCutQ = -1.0f;
 
 	JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(DelayEngine)
 };